09. Filtering, Sorting, and Pagination | 11. Refactoring. Extracting IssueTable Component
