+++ proj/src/app/issues/IssueActions.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 
 const IssueActions = () => {
   return (
-    <Flex mb="5" justify="between">
+    <Flex justify="between">
       <IssueStatusFilter />
       <Button>
         <Link href="/issues/new">New Issue</Link>
